IDE-1957 Correct issues with ScaUriHelpers
If the getLocalFileResource() method was invoked twice with the same arguments,
and the request was for a non-existent file (or really just any file that has
issues and won't load), the method returned an empty Resource with errors on
the 2nd request (and all subsequent ones).

The issue was that when loading a Resource fails, a Resource still gets added
into the ResourceSet. The code in getLocalFileEObject() was structured such
that the Resource is never returned, but getLocalFileResource() had a bug.