Add new CORBA helpers
The older helpers wrap any exception in a CoreException and throw. However,
it is common for us to throw a CoreException within the Callable; those
exceptions are re-thrown as-is. This makes it hard to know if the CoreException
you're seeing was thrown to wrap an unanticipated exception, or if it was
thrown by the code in the Callable.

The new method adds a few improvements:
* Translates any kind of cancellation into an OperationCancelledException,
  which is advantageous within Jobs and other Eclipse methods
* Does not throw InterruptedException (thus client code doesn't have to deal
  with it), but also re-interrupts the thread so the interrupt isn't lost.
* Uses ExecutionException to wrap exceptions, preventing confusion with
  CoreException.

A second method is also added that uses InvocationTargetException instead of
ExecutionException, which is useful when called within an IRunnableWithProgress
in the UI.